TINY-9003: Refactoring and fixes from review
