+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - Enabled the `@typescript-eslint/naming-convention` rule so that it enforces pascal case type names.
-- Added new `@tinymce/prefer-mcagar` rule to detect places where `TinyDom`, `TinyAssertions.assertContent` and `TinyAssertions.assertRawContent` should be used.
+- Added new `@tinymce/prefer-mcagar-tiny-assertions` rule to detect places where `TinyAssertions.assertContent` and `TinyAssertions.assertRawContent` should be used in tests.
+- Added new `@tinymce/prefer-mcagar-tiny-dom` rule to detect places where `TinyDom` should be used in tests.
 - Added new `@tinymce/no-direct-editor-events` rule to ensure events are only dispatched inside `api/Events.ts`.
 - Added new `@tinymce/no-direct-editor-options` rule to ensure options are only registered and accessed inside `api/Options.ts`.
 - Added new `@tinymce/no-publicapi-module-imports` rule to detect invalid imports from `PublicApi.ts`.
