+++ src/main/ts/configs/Standard.ts
@@ -169,6 +169,7 @@ export const standard: Linter.Config = {
     '@tinymce/no-unimported-promise': 'off',
     '@tinymce/no-implicit-dom-globals': 'error',
     '@tinymce/prefer-fun': 'error',
-    '@tinymce/prefer-mcagar': 'off',
+    '@tinymce/prefer-mcagar-tiny-assertions': 'off',
+    '@tinymce/prefer-mcagar-tiny-dom': 'off',
   }
 };
